pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cephla_tilefusion"
version = "0.1.1"
description = "GPU/CPU-accelerated tile registration and fusion for 2D microscopy"
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.9"
authors = [
    { name = "Cephla" },
]
keywords = [
    "microscopy",
    "image-stitching",
    "tile-fusion",
    "ome-tiff",
    "zarr",
    "gpu-acceleration",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.21",
    "tifffile>=2023.0",
    "tensorstore>=0.1.45",
    "numba>=0.57",
    "pandas>=1.5",
    "scikit-image>=0.21",
    "scipy>=1.10",
    "tqdm>=4.65",
    "psutil>=5.9",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=12.0",
    "cucim>=23.0",
]
gui = [
    "PyQt5>=5.15",
    "napari>=0.4.18",
    "ome-zarr>=0.8",
]
flatfield = [
    "basicpy @ git+https://github.com/QI2lab/BaSiCPy.git",
    "matplotlib>=3.5",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = [
    "cephla_tilefusion[gpu,gui,flatfield,dev]",
]

[project.scripts]
stitcher-gui = "gui.app:main"
convert-to-zarr = "scripts.convert_to_zarr:main"

[project.urls]
Homepage = "https://github.com/cephla/tilefusion"
Documentation = "https://github.com/cephla/tilefusion#readme"
Repository = "https://github.com/cephla/tilefusion"
Issues = "https://github.com/cephla/tilefusion/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cephla_tilefusion", "gui", "scripts"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=cephla_tilefusion --cov-report=term-missing"
